split opt64 impl. and SIMD dispatch; implement plain64 build
